refactor: separate library core from CLI adapter (closes #62)
- Rename src/agent/ → src/core/ (agent.ts, executor.ts, context.ts, prompt.ts)
  - src/agent/core.ts → src/core/agent.ts (avoids redundant name)
- Rename src/model/ → src/providers/ (types, client, gemini, anthropic, openai, factory, schema)
- factory.ts: remove Config dependency; createClient(model, apiKey) now accepts
  an explicit API key instead of reading process.env or config internally
- cli/index.ts: add resolveApiKey() that reads env vars + config and passes
  the resolved key to createClient() — all env/config I/O stays in the CLI layer
- Update all import paths across core/, tools/, state/, cli/
- Update CLAUDE.md and AGENTS.md source structure and architecture descriptions

The rule: core/ and providers/ import nothing from cli/ or state/.
Programmatic usage is now possible without importing CLI modules.
