+++ src/cli/repl.ts
@@ -2,12 +2,12 @@ import chalk from "chalk";
 import { execFileSync } from "node:child_process";
 import { writeFile, readFile, rm } from "node:fs/promises";
 import { tmpdir } from "node:os";
-import type { Agent, AgentRunMode } from "../agent/core.js";
+import type { Agent, AgentRunMode } from "../core/agent.js";
 import type { SkillRegistry } from "../skills/registry.js";
 import { loadSkillFile, processBody } from "../skills/loader.js";
 import { join } from "node:path";
 import { readLine, selectKey, loadHistory, saveHistory, type SlashCommand } from "./input.js";
-import type { ConfirmFn } from "../agent/executor.js";
+import type { ConfirmFn } from "../core/executor.js";
 import { loadConfig, saveConfig } from "../state/config.js";
 import { loadSettings, saveSettings } from "../state/settings.js";
 import { Session } from "../state/session.js";
