+++ src/core/prompt.test.ts
