+++ src/providers/factory.ts
@@ -2,7 +2,6 @@ import { GeminiClient } from "./gemini.js";
 import { AnthropicClient } from "./anthropic.js";
 import { OpenAIClient } from "./openai.js";
 import type { LLMClient } from "./client.js";
-import type { Config } from "../state/config.js";
 
 export type Provider = "gemini" | "anthropic" | "openai";
 
@@ -30,34 +29,9 @@ export function hasNativeThinking(model: string): boolean {
   return false;
 }
 
-export function createClient(model: string, config: Config): LLMClient {
+export function createClient(model: string, apiKey: string): LLMClient {
   const provider = detectProvider(model);
-
-  if (provider === "anthropic") {
-    const key = process.env.ANTHROPIC_API_KEY ?? config.anthropicApiKey;
-    if (!key) {
-      throw new Error(
-        "No Anthropic API key found. Set ANTHROPIC_API_KEY or run: opencli config --anthropic-api-key <key>",
-      );
-    }
-    return new AnthropicClient(key, model);
-  }
-
-  if (provider === "openai") {
-    const key = process.env.OPENAI_API_KEY ?? config.openaiApiKey;
-    if (!key) {
-      throw new Error(
-        "No OpenAI API key found. Set OPENAI_API_KEY or run: opencli config --openai-api-key <key>",
-      );
-    }
-    return new OpenAIClient(key, model);
-  }
-
-  const key = process.env.GEMINI_API_KEY ?? config.geminiApiKey;
-  if (!key) {
-    throw new Error(
-      "No Gemini API key found. Set GEMINI_API_KEY or run: opencli config --gemini-api-key <key>",
-    );
-  }
-  return new GeminiClient(key, model);
+  if (provider === "anthropic") return new AnthropicClient(apiKey, model);
+  if (provider === "openai") return new OpenAIClient(apiKey, model);
+  return new GeminiClient(apiKey, model);
 }
